+++ pizzaordersystem/src/main/resources/messages_en.properties
@@ -139,10 +139,10 @@ customer.details.modal.update.button.submit.label=Submit
 employee.customer.list.table.heading.id=Customer Id
 employee.customer.list.table.heading.name=Customer Name
 employee.customer.list.table.heading.address.line.1=Address Line 1
-employee.customer.list.table.heading.adddress.line.2=Address Line 2
+employee.customer.list.table.heading.address.line.2=Address Line 2
 employee.customer.list.table.heading.city=City Name
 employee.customer.list.table.heading.state=State Name
-employee.customer.list.table.heading.country.name=Country Name
+employee.customer.list.table.heading.country=Country Name
 employee.customer.list.table.heading.gender=Gender
 employee.customer.list.table.heading.email=Email
 employee.customer.list.table.heading.phone=Phone Number
