+++ pizzaordersystem/src/main/webapp/WEB-INF/views/customerdetails.jsp
@@ -162,6 +162,8 @@
 	<script type="text/javascript"
 		src="<c:url value="/static/js/login.js" />"></script>
 	<script type="text/javascript"
+		src="<c:url value="/static/js/validation.js" />"></script>
+	<script type="text/javascript"
 		src="<c:url value="/static/js/customer-details.js" />"></script>
 </body>
 </html>
